+++ source/extensions/lightspeed.trex.project_wizard.core/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## [1.1.1] - 2023-09-06
+### Fixed
+- Fixed creating a project with no capture selected
+
 ## [1.1.0] - 2023-09-05
 ### Changed
 - Added extra checks to `is_project_file_valid`